// src/lib/api/generated/model/v1SandboxTemplateStatus.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Aisphere Hub API
 * OpenAPI spec version: 1.0.0
 */

export type V1SandboxTemplateStatus = typeof V1SandboxTemplateStatus[keyof typeof V1SandboxTemplateStatus];


export const V1SandboxTemplateStatus = {
  SANDBOX_TEMPLATE_STATUS_UNSPECIFIED: 'SANDBOX_TEMPLATE_STATUS_UNSPECIFIED',
  SANDBOX_TEMPLATE_STATUS_CREATING: 'SANDBOX_TEMPLATE_STATUS_CREATING',
  SANDBOX_TEMPLATE_STATUS_READY: 'SANDBOX_TEMPLATE_STATUS_READY',
  SANDBOX_TEMPLATE_STATUS_FAILED: 'SANDBOX_TEMPLATE_STATUS_FAILED',
  SANDBOX_TEMPLATE_STATUS_DELETED: 'SANDBOX_TEMPLATE_STATUS_DELETED',
} as const;